+++ bgstally/windows/colonisation.py
@@ -166,7 +166,7 @@ class ColonisationWindow:
         self.legend_fr:tk.Toplevel|None = None
         self.notes_fr:tk.Toplevel|None = None
         self.bases_fr:tk.Toplevel|None = None
-        self.bodies_popup:BodiesPopup = None
+        self.bodies_fr:tk.Toplevel = None
         self.scale:float = 0
 
 
@@ -184,7 +184,8 @@ class ColonisationWindow:
         self.window.title(_("{plugin_name} - Colonisation").format(plugin_name=self.bgstally.plugin_name)) # LANG: window title
 
         self.window.minsize(400, 100)
-        self.window.geometry(f"{int(1500*self.scale)}x{int(500*self.scale)}")
+        geometry:str = self.colonisation.window_geometries.get('Colonisation', f"{int(1500*self.scale)}x{int(500*self.scale)}")
+        self.window.geometry(geometry)
         self.window.protocol("WM_DELETE_WINDOW", self.close)
         self._create_frames()   # Create main frames
         self.update_display()   # Populate them
@@ -216,7 +217,7 @@ class ColonisationWindow:
         if tabnum > 1:
             t:int = 1
             for t in range(1, tabnum-1):
-                Debug.logger.debug(f"{t} {self.tl.get(t)} {tabnum-1}")
+                #Debug.logger.debug(f"{t} {self.tl.get(t)} {tabnum-1}")
                 if systems[self.tl[t]].get('Hidden', True) == False:
                     break
             self.tabbar.select(t) # Select the first non-hidden system tab
@@ -393,10 +394,10 @@ class ColonisationWindow:
         if dest == '' or dest == None:
             dest = config.get_str('system_provider')
         if type not in self.links.keys():
-            Debug.logger.debug(f"Unknown link type: {type}")
+            Debug.logger.info(f"Unknown link type: {type}")
             return
         if dest not in self.links[type].keys():
-            Debug.logger.debug(f"Unknown destination: {dest}")
+            Debug.logger.info(f"Unknown destination: {dest}")
             return
         params:dict = {k: quote(str(v)) if str(k) != 'Layout' else str(v).strip().lower().replace(" ","_") for k, v in data.items()}
         webbrowser.open(self.links[type][dest].format(**params))
@@ -420,8 +421,9 @@ class ColonisationWindow:
 
         self.bases_fr = tk.Toplevel(self.bgstally.ui.frame)
         self.bases_fr.wm_title(_("{plugin_name} - Colonisation Base Types").format(plugin_name=self.bgstally.plugin_name)) # LANG: Title of the base type popup window
-        self.bases_fr.geometry(f"{int(1000*self.scale)}x{int(500*self.scale)}")
-        self.bases_fr.protocol("WM_DELETE_WINDOW", self.bases_fr.destroy)
+        geometry:str = self.colonisation.window_geometries.get('Bases', f"{int(1000*self.scale)}x{int(500*self.scale)}")
+        self.bases_fr.geometry(geometry)
+        self.bases_fr.protocol("WM_DELETE_WINDOW", partial(self.close, 'Bases', self.bases_fr))
         self.bases_fr.config(bd=2, relief=tk.FLAT)
         header_fnt:tuple = (FONT_SMALL[0], FONT_SMALL[1], "bold")
         sheet:Sheet = Sheet(self.bases_fr, sort_key=natural_sort_key, note_corners=True, show_row_index=False, cell_auto_resize_enabled=True, height=4096,
@@ -478,8 +480,7 @@ class ColonisationWindow:
 
             notes:str = text.get("1.0", tk.END)
             self.colonisation.modify_system(system, {'Notes': notes.strip()})
-            self.notes_fr.destroy()
-            self.notes_fr
+            self.close('Notes', self.notes_fr)
 
         sysnum:int = self.tl[tabnum]
         systems:list = self.colonisation.get_all_systems()
@@ -490,7 +491,9 @@ class ColonisationWindow:
         self.notes_fr = tk.Toplevel(self.bgstally.ui.frame)
         self.notes_fr.wm_title(_("{plugin_name} - Colonisation Notes for {system_name}").format(plugin_name=self.bgstally.plugin_name, system_name=systems[sysnum].get('Name', ''))) # LANG: Title of the notes popup window
         self.notes_fr.wm_attributes('-topmost', True)     # keeps popup above everything until closed.
-        self.notes_fr.geometry("600x600")
+        geometry:str = self.colonisation.window_geometries.get('Notes', "600x600")
+        self.notes_fr.geometry(geometry)
+        self.notes_fr.protocol("WM_DELETE_WINDOW", partial(self.close, 'Notes', self.notes_fr))
         self.notes_fr.config(bd=2, relief=tk.FLAT)
         scr:tk.Scrollbar = tk.Scrollbar(self.notes_fr, orient=tk.VERTICAL)
         scr.pack(side=tk.RIGHT, fill=tk.Y)
@@ -512,7 +515,9 @@ class ColonisationWindow:
         self.legend_fr = tk.Toplevel(self.bgstally.ui.frame)
         self.legend_fr.wm_title(_("{plugin_name} - Colonisation Legend").format(plugin_name=self.bgstally.plugin_name)) # LANG: Title of the legend popup window
         self.legend_fr.wm_attributes('-topmost', True)     # keeps popup above everything until closed.
-        self.legend_fr.geometry(f"600x600")
+        geometry:str = self.colonisation.window_geometries.get('Legend', f"600x600")
+        self.legend_fr.geometry(geometry)
+        self.legend_fr.protocol("WM_DELETE_WINDOW", partial(self.close, 'Legend', self.legend_fr))
         self.legend_fr.config(bd=2, relief=tk.FLAT)
 
         text:RichScrolledText = RichScrolledText(self.legend_fr, markdown=self._load_legend())
@@ -1275,7 +1280,9 @@ class ColonisationWindow:
 
         dialog.title(_("Edit System")) # LANG: Rename a system
         dialog.minsize(500, 250)
-        dialog.geometry(f"{int(500*self.scale)}x{int(250*self.scale)}+{btn.winfo_x()-100}+{self.window.winfo_y()+50}")
+        geometry:str = self.colonisation.window_geometries.get('EditSystem', f"{int(500*self.scale)}x{int(250*self.scale)}+{btn.winfo_x()-100}+{self.window.winfo_y()+50}")
+        dialog.geometry(geometry)
+        dialog.protocol("WM_DELETE_WINDOW", partial(self.close, 'EditSystem', dialog))
         #dialog.transient(self.window)
         dialog.config(bd=2, relief=tk.FLAT)
         dialog.grab_set()
@@ -1324,7 +1331,7 @@ class ColonisationWindow:
         cancel_button:ttk.Button = ttk.Button(
             button_frame,
             text=_("Cancel"), # LANG: Cancel button
-            command=dialog.destroy
+            command=partial(self.close, 'EditSystem', dialog)
         ) # LANG: Cancel button
         cancel_button.pack(side=tk.LEFT, padx=5)
 
@@ -1354,7 +1361,7 @@ class ColonisationWindow:
         self.tabbar.notebookTab.tab(tabnum, text=name)
         self.update_react_dialog()
 
-        dialog.destroy()
+        self.close('EditSystem', dialog)
         self.update_display()
 
 
@@ -1411,13 +1418,23 @@ class ColonisationWindow:
 
 
     @catch_exceptions
-    def close(self) -> None:
+    def close(self, n:str = '', w:tk.Toplevel|None = None) -> None:
         ''' Close the window and any popups and clean up'''
-        if self.window: self.window.destroy()
-        if self.legend_fr: self.legend_fr.destroy()
-        if self.notes_fr: self.notes_fr.destroy()
-        if self.bases_fr: self.bases_fr.destroy()
-        if self.bodies_popup and self.bodies_popup.window: self.bodies_popup.window.destroy()
+        # Close one window.
+        if w and w.winfo_exists():
+            self.colonisation.window_geometries[n] = w.winfo_geometry()
+            w.destroy()
+            return
+
+        # Save geometry nd close all windows
+        for n, w in {'Bodies' : self.bodies_fr,
+                     'Bases' : self.bases_fr,
+                     'Notes' : self.notes_fr,
+                     'Legend' : self.legend_fr,
+                     'Colonisation' : self.window}.items():
+            if w and w.winfo_exists():
+                self.colonisation.window_geometries[n] = w.winfo_geometry()
+                w.destroy()
 
         # UI components
         self.tabbar = None
@@ -1596,8 +1613,12 @@ class BodiesPopup:
         self.window:tk.Toplevel = tk.Toplevel(parent.bgstally.ui.frame)
         self.window.wm_title(_("{plugin_name} - Colonisation Bodies").format(plugin_name=parent.bgstally.plugin_name)) # LANG: Title of the bodies popup window
         self.window.minsize(800, 600)
-        self.window.geometry(f"{int(800*parent.scale)}x{int(600*parent.scale)}")
+        geometry:str = parent.colonisation.window_geometries.get('Bodies', f"{int(800*parent.scale)}x{int(600*parent.scale)}")
+        self.window.geometry(geometry)
         self.window.config(bd=2, relief=tk.FLAT)
+        # This handles geometry saving
+        parent.bodies_fr = self.window
+        self.window.protocol("WM_DELETE_WINDOW", partial(parent.close, 'Bodies', self.window))
 
         self.images:dict = {}
 
@@ -1680,10 +1701,10 @@ class BodiesPopup:
         ''' Recursive function to walk the body list and add items to the treeview.
             It assumes the bodies are in order of parent-child relationships and that the 'parents' attribute indicates the level of the body in the hierarchy. '''
 
-        Debug.logger.debug(f"Walking tree with parent_id {parent_id} and parent_node {parent_node}")
+        #Debug.logger.debug(f"Walking tree with parent_id {parent_id} and parent_node {parent_node}")
         for i, body in enumerate(bodies):
             if body.get('parentId', 0) == parent_id:
-                Debug.logger.warning(f"Inserting {body.get('name', '')} ({body.get('parentId', -1)}) below {parent_id} ({parent_node})")
+                #Debug.logger.warning(f"Inserting {body.get('name', '')} ({body.get('parentId', -1)}) below {parent_id} ({parent_node})")
 
                 name:str = body.get('name', '')
                 if name.startswith(system.get('StarSystem', '')) and body.get('type', '') != 'Star':
@@ -1775,7 +1796,7 @@ class BodiesPopup:
 
         features:str = ''
         flist:list = []
-        Debug.logger.debug(f"Calculating features for {body.get('name', '')} with features {body.get('features', [])}")
+        #Debug.logger.debug(f"Calculating features for {body.get('name', '')} with features {body.get('features', [])}")
         if 'features' in body and len(body['features']) > 0:
             flist = [f[0].upper() for f in body['features'] if f not in ['Terraformable', 'Landable']]
             features = " ".join(flist)
@@ -1790,7 +1811,7 @@ class BodiesPopup:
                 bt:dict = self.parent.colonisation.get_base_type(b.get('Base Type', ''))
                 bt['State'] = b.get('State', BuildState.PLANNED)
                 if i == 0: bt['Primary'] = True # Mark the first build as primary
-                Debug.logger.debug(f"Adding build {b.get('Name', '')} to body {body.get('name', '')} {b}")
+                #Debug.logger.debug(f"Adding build {b.get('Name', '')} to body {body.get('name', '')} {b}")
                 build = tree.insert(item, 'end', image=self._get_body_icon(bt), text=" " +b.get('Base Type', ''), \
                                     values=(b.get('BuildID'), b.get('Name'), '', '', ''), tags=('base',))
                 tree.item(build, tags=('base',))