+++ bgstally/windows/fleetcarrier.py
@@ -3,7 +3,7 @@ from functools import partial
 from tkinter import ttk
 
 #from bgstally.bgstally import BGSTally
-from bgstally.constants import DATETIME_FORMAT_CARRIER, FONT_SMALL, COLOUR_WARNING, DiscordChannel, DiscordFleetCarrier
+from bgstally.constants import DATETIME_FORMAT_CARRIER, FONT_SMALL, FONT_HEADING_1, COLOUR_WARNING, DiscordChannel, DiscordFleetCarrier
 from bgstally.fleetcarrier import FleetCarrier
 from bgstally.utils import _, __, hfplus, str_truncate, catch_exceptions
 from bgstally.widgets import TreeviewPlus, AutoCompleter, Placeholder
@@ -26,6 +26,8 @@ class WindowFleetCarrier:
         self.window:tk.Toplevel|None = None
         self.frame:ttk.Frame
         self.itineraryfr:ttk.Frame
+        self.summfr:ttk.Frame|None = None
+        self.tabbar:ScrollableNotebook|None = None
         self.scale:float = 1.0
 
         self.tabs:dict = {
@@ -112,6 +114,7 @@ class WindowFleetCarrier:
             },
         }
 
+        #self.cooldown_notice()
 
     @catch_exceptions
     def show(self) -> None:
@@ -124,7 +127,9 @@ class WindowFleetCarrier:
         self.window = tk.Toplevel(self.bgstally.ui.frame)
         self.window.title(_("{plugin_name} - Carrier {carrier_name}").format(plugin_name=self.bgstally.plugin_name, carrier_name=self.bgstally.fleet_carrier.overview.get('name'))) # LANG: Carrier window title
         self.window.iconphoto(False, self.bgstally.ui.image_logo_bgstally_32, self.bgstally.ui.image_logo_bgstally_16)
-        self.window.geometry(f"{int(850*self.scale)}x{int(550*self.scale)}")
+        geometry:str = self.bgstally.fleet_carrier.window_geometries.get('Carrier', f"{int(850*self.scale)}x{int(550*self.scale)}")
+        self.window.geometry(geometry)
+        self.window.protocol("WM_DELETE_WINDOW", self.close)
 
         self.frame = ttk.Frame(self.window)
         if not config.get_bool('capi_fleetcarrier'):
@@ -139,34 +144,60 @@ class WindowFleetCarrier:
         if self.window == None or not self.window.winfo_exists(): return
 
         # Clear existing contents
-        for w in self.frame.winfo_children(): w.destroy()
+        #for w in self.frame.winfo_children(): w.destroy()
 
         self._show_overview(self.bgstally.fleet_carrier, self.frame)
         self._create_tabs(self.bgstally.fleet_carrier, self.frame)
 
+    def close(self, n:str = '', w:tk.Toplevel|None = None) -> None:
+        ''' Close the window and any popups and clean up'''
+        # Close one window.
+        if w and w.winfo_exists():
+            self.bgstally.fleet_carrier.window_geometries['Alert'] = w.winfo_geometry()
+            w.destroy()
+            return
+
+        if self.window and self.window.winfo_exists():
+            self.bgstally.fleet_carrier.window_geometries['Carrier'] = self.window.winfo_geometry()
+            self.window.destroy()
+
+        # UI components
+        self.summfr = None
+        self.tabbar = None
+        for k, v in self.tabs.items():
+            v['fr'] = None
+
 
     def _show_overview(self, fc:FleetCarrier, frame:ttk.Frame) -> None:
         """ Show the Fleet Carrier overview tab """
-        summ:ttk.Frame = ttk.Frame(frame)
-        summ.pack(fill=tk.X)
-        self._create_columns(fc.get_overview(), 3, summ)
+        if self.summfr == None:
+            self.summfr = ttk.Frame(frame)
+            self.summfr.pack(fill=tk.X)
+        self._create_columns(fc.get_overview(), 3, self.summfr)
 
 
     def _create_tabs(self, fc:FleetCarrier, frame:ttk.Frame) -> None:
         """ Create and populate the Fleet Carrier tabs """
 
-        style:ttk.Style = ttk.Style()
-        style.configure("White.TNotebook.Tab", font=(FONT_SMALL[0], FONT_SMALL[1], "bold"), padding=[10, 5], background='green')
-        tabbar:ScrollableNotebook = ScrollableNotebook(frame, wheelscroll=True, tabmenu=False, style='White.TNotebook')
-        tabbar.pack(fill=tk.X, padx=5, pady=5)
+        # Only create this once.
+        if self.tabbar == None:
+            style:ttk.Style = ttk.Style()
+            style.configure("White.TNotebook.Tab", font=(FONT_SMALL[0], FONT_SMALL[1], "bold"), padding=[10, 5], background='green')
+            self.tabbar = ScrollableNotebook(frame, wheelscroll=True, tabmenu=False, style='White.TNotebook')
+            self.tabbar.pack(fill=tk.X, padx=5, pady=5)
 
         for k, v in self.tabs.items():
-            fr:ttk.Frame = ttk.Frame(tabbar, relief=tk.FLAT)
-            fr.pack(fill=tk.BOTH, expand=1)
-            tabbar.add(fr, text=_(k)) # LANG: Ignore
+            # Only create these once
+            if v.get('fr', None) == None:
+                v['fr'] = ttk.Frame(self.tabbar, relief=tk.FLAT)
+                v['fr'].pack(fill=tk.BOTH, expand=1)
+                self.tabbar.add(v['fr'], text=_(k)) # LANG: Ignore
+
+            # Create/Recreate the contents of each tab
+            for w in v['fr'].winfo_children(): w.destroy()
             if v.get('buttons', None) != None:
-                v['buttons'](fc, fr)
-            v['func'](fc, v, fr)
+                v['buttons'](fc, v['fr'])
+            v['func'](fc, v, v['fr'])
 
 
     def _summary(self, fc:FleetCarrier, which:dict, frame:ttk.Frame) -> None:
@@ -315,7 +346,7 @@ class WindowFleetCarrier:
         ''' Create grid of title/value pairs in columns and rows '''
         row:int = 0; col = 0
         lbl:ttk.Label
-
+        for w in frame.winfo_children(): w.destroy()
         for k, v in data.items():
             if bg != 'None':
                 lbl = ttk.Label(frame, text=_(k), font=(FONT_SMALL[0], FONT_SMALL[1], "bold"), background=bg) # LANG: Ignore
@@ -534,3 +565,44 @@ class WindowFleetCarrier:
                 output += " | ".join(line) + "\n"
         output += "```"
         return output
+
+
+    def cooldown_notice(self) -> None:
+        """ Display carrier cooldown notification """
+        self.bgstally.ui.show_warning("Fleetcarrier cooldown\ncompleted")
+        PopupNotice(_("Fleetcarrier cooldown\ncompleted"), 20000, self.bgstally.fleet_carrier)
+
+class PopupNotice:
+    """ Create a temporary popup window """
+    def __init__(self, notice:str = '', timeout:int = 0, fc:FleetCarrier|None = None) -> None:
+        self.fc:FleetCarrier|None = fc
+        self.root = tk.Tk()
+        self.root.overrideredirect(True)
+        self.root.attributes("-alpha", 0.6)
+        self.root.geometry(fc.window_geometries.get('Alert', "300x150-1+0"))
+        self.root.attributes("-topmost", True)
+        self.frame = tk.Frame(self.root, bg='red4', relief="raised")
+        self.frame.pack(fill="both", expand=True)
+        label = tk.Label(self.frame, text=notice, fg="white", bg="red4", font=FONT_HEADING_1, justify=tk.CENTER)
+        label.pack(pady=20, anchor=tk.CENTER)
+        exit_btn = tk.Button(self.frame, text=_("Close"), fg="white", bg="red4", command=self.close)
+        exit_btn.pack(pady=10)
+        if timeout > 0: self.root.after(timeout, self.close)
+        self.frame.bind("<Button-1>", self.start_move)
+        self.frame.bind("<B1-Motion>", self.do_move)
+
+    def start_move(self, event) -> None:
+        self.x:int = event.x
+        self.y:int = event.y
+
+    def do_move(self, event) -> None:
+        deltax:int = event.x - self.x
+        deltay:int = event.y - self.y
+        x:int = self.root.winfo_x() + deltax
+        y:int = self.root.winfo_y() + deltay
+        self.root.geometry(f"+{x}+{y}")
+
+    def close(self) -> None:
+        if self.root and self.root.winfo_exists():
+            self.fc.window_geometries['Alert'] = self.root.winfo_geometry()
+            self.root.destroy()